Move applications getters from utils to model
